Homepage Factories Tests
1. Move homepage app tests into a tests directory
2. Refactor the HomepageFactory, HomepageLeadersFactory, HomepageNewsFactory and HomepageHighlightsFactory tests so they test only what makes sense - i.e We shouldn't test that functionality provided by FakerBoy can pass arguments to sub-factories
